separated unit test file into dif files by module
