+++ src/tester/test_unittest_scraper.py
@@ -25,14 +25,12 @@ project_root = os.path.dirname(parent_dir)
 
 SKIP_SLOW = os.getenv("SKIP_SLOW", "false").lower().strip() == "true"
 
-
 def log(
     message, level="INFO"
 ):  # Provide message and info level (optional, defaulting to info)
     # configure the logger
     log = logging.getLogger(__name__)
     log.info(message)
-
 
 class ScraperTestCase(unittest.TestCase):
     # Defaults for each program are set at the function level.
@@ -783,385 +781,3 @@ class ScraperTestCase(unittest.TestCase):
             "The cause number is not where it was expected to be in the HTML.",
         )
         # self.logger.info(f"Scraper test sucessful for cause number CR-16-0002-A.")"""
-
-
-class ParseTestCase(unittest.TestCase):
-    def setUp(self):
-        self.test_dir = tempfile.mkdtemp()
-        self.case_json_path = os.path.join(self.test_dir, "hays", "case_json")
-        os.makedirs(self.case_json_path, exist_ok=True)
-
-        self.mock_logger = logging.getLogger(__name__)
-        self.parser_instance = parser.Parser()
-        self.case_html_path = os.path.abspath(
-            os.path.join(
-                os.path.dirname(__file__), "../../resources/test_files/parser_testing"
-            )
-        )
-
-    def test_parser_class_and_method(self):
-        parser_instance = parser.Parser()
-
-        instance, method = parser_instance.get_class_and_method(
-            logger=self.mock_logger, county="hays", test=True
-        )
-        self.assertIn('extract_rows', dir(instance))
-
-    @patch("os.makedirs")
-    def test_parser_directories_single_file(self, mock_makedirs):
-        parser_instance = parser.Parser()
-        case_html_path, case_json_path = parser_instance.get_directories(
-            "hays", self.mock_logger, parse_single_file=True
-        )
-
-        base_dir = os.path.abspath(os.path.join(os.path.dirname(__file__), "..", ".."))
-        expected_path = os.path.join(base_dir, "resources", "test_files")
-
-        self.assertEqual(case_html_path, expected_path)
-        self.assertEqual(case_json_path, expected_path)
-
-    @patch("os.makedirs")
-    @patch("os.path.exists", return_value=False)
-    def test_parser_directories_multiple_files(self, mock_exists, mock_makedirs):
-        parser_instance = parser.Parser()
-        case_html_path, case_json_path = parser_instance.get_directories(
-            "hays", self.mock_logger, parse_single_file=False
-        )
-
-        base_dir = os.path.abspath(os.path.join(os.path.dirname(__file__), "..", ".."))
-        expected_html_path = os.path.join(base_dir, "data", "hays", "case_html")
-        expected_json_path = os.path.join(base_dir, "data", "hays", "case_json")
-
-        self.assertEqual(case_html_path, expected_html_path)
-        self.assertEqual(case_json_path, expected_json_path)
-        mock_makedirs.assert_called_once_with(expected_json_path, exist_ok=True)
-
-    def test_parser_list_of_single_html_file(self):
-        case_number = "51652356"
-        case_list = self.parser_instance.get_list_of_html(
-            self.case_html_path,
-            case_number,
-            "hays",
-            self.mock_logger,
-            parse_single_file=True,
-        )
-
-        relative_path = os.path.join(project_root, "resources", "test_files")
-
-        expected_path = os.path.join(relative_path, f"test_{case_number}.html")
-
-        self.assertEqual(case_list, [expected_path])
-
-    def test_parser_list_of_single_html_file_by_casenumber(self):
-        case_number = "51652356"
-
-        case_list = self.parser_instance.get_list_of_html(
-            self.case_html_path,
-            case_number,
-            "hays",
-            self.mock_logger,
-            parse_single_file=True,
-        )
-
-        relative_path = os.path.join(project_root, "resources", "test_files")
-
-        expected_list = [os.path.join(relative_path, f"test_{case_number}.html")]
-
-        self.assertEqual(case_list, expected_list)
-
-    def test_parser_list_of_multiple_html_files(self):
-        os.makedirs(self.case_html_path, exist_ok=True)
-
-        with open(os.path.join(self.case_html_path, "test_1.html"), "w") as f:
-            f.write("test")
-        with open(os.path.join(self.case_html_path, "test_2.html"), "w") as f:
-            f.write("test")
-
-        updated_html_path = os.path.join(self.case_html_path, "multiple_html_files")
-        case_number = ""
-        case_list = self.parser_instance.get_list_of_html(
-            updated_html_path,
-            case_number,
-            "hays",
-            self.mock_logger,
-            parse_single_file=False,
-        )
-
-        expected_list = [
-            os.path.join(updated_html_path, "test_1.html"),
-            os.path.join(updated_html_path, "test_2.html"),
-        ]
-
-        self.assertEqual(set(case_list), set(expected_list))
-
-    def test_parser_get_list_of_html_error_handling(self):
-        invalid_path = "invalid/path"
-        case_number = "12345"
-
-        with self.assertRaises(Exception):
-            self.parser_instance.get_list_of_html(
-                invalid_path,
-                case_number,
-                "hays",
-                self.mock_logger,
-                parse_single_file=False,
-            )
-
-    def test_get_html_path(self):
-        updated_html_path = os.path.join(self.case_html_path, "multiple_html_files")
-        case_html_file_name = "parserTest_51652356.html"
-        case_number = "51652356"
-
-        result = self.parser_instance.get_html_path(
-            updated_html_path, case_html_file_name, case_number, self.mock_logger
-        )
-
-        self.assertEqual(result, f"{os.path.join(updated_html_path,case_html_file_name)}")
-
-    @patch("builtins.open", new_callable=mock_open)
-    def test_write_json_data(self, mock_open_func):
-        case_json_path = "/mock/path"
-        case_number = "123456"
-        case_data = {"data": "value"}
-
-        self.parser_instance.write_json_data(
-            case_json_path, case_number, case_data, self.mock_logger
-        )
-
-        mock_open_func.assert_called_once_with(
-            os.path.join(case_json_path, case_number + ".json"), "w"
-        )
-
-    @patch("builtins.open", new_callable=mock_open)
-    def test_write_error_log(self, mock_open_func):
-        county = "hays"
-        case_number = "123456"
-
-        self.parser_instance.write_error_log(county, case_number)
-
-        base_dir = os.path.abspath(os.path.join(os.path.dirname(__file__), "..", ".."))
-        error_log_path = os.path.join(
-            base_dir, "data", county, "cases_with_parsing_error.txt"
-        )
-
-        mock_open_func.assert_called_once_with(error_log_path, "w")
-
-    def test_parser_end_to_end(self, county="hays", case_number='123456'):
-
-        self.parser_instance.parse(county=county, 
-                     case_number=case_number, 
-                     parse_single_file=True,
-                     test = True)
-
-class CleanTestCase(unittest.TestCase):
-    def setUp(self):
-        self.cleaner = cleaner.Cleaner() # Create Cleaner instance here to avoid repeating this in every test
-
-    @patch('os.makedirs') 
-    @patch('os.path.exists', return_value=False)
-    def test_get_or_create_folder_path(self, mock_exists, mock_makedirs):
-        mock_exists.return_value = False
-        county = "hays"
-        folder_type = "case_json"
-        cleaner_dir = os.path.abspath(os.path.join(os.path.dirname(__file__), "..", "cleaner"))
-        expected_path = os.path.join(cleaner_dir, "..", "..", "data", county, folder_type)
-
-        folder_path = self.cleaner.get_or_create_folder_path(county, folder_type)
-
-        mock_exists.assert_called_once_with(expected_path)  # Check if os.path.exists was called
-        mock_makedirs.assert_called_once_with(expected_path)  # Check if os.makedirs was called
-        self.assertEqual(folder_path, expected_path)  # Check that the path is correct
-
-        # Test when folder already exists
-        mock_exists.return_value = True
-        folder_path = self.cleaner.get_or_create_folder_path(county, folder_type)
-        mock_makedirs.assert_called_once() # Should not be called again
-
-    def test_load_json_file(self):
-        # Test successful load
-        with patch("builtins.open", new_callable=mock_open, read_data='{"key": "value"}'):
-            result = self.cleaner.load_json_file("fake_path.json")
-            self.assertEqual(result, {"key": "value"})
-
-        # Test file not found
-        with patch("builtins.open", side_effect=FileNotFoundError):
-            result = self.cleaner.load_json_file("nonexistent.json")
-            self.assertEqual(result, {})
-        
-        # Test invalid JSON
-        with patch("builtins.open", new_callable=mock_open, read_data='invalid json'):
-            result = self.cleaner.load_json_file("invalid.json")
-            self.assertEqual(result, {})
-
-    def test_load_and_map_charge_names(self):
-        # Test successful mapping
-        test_data = '[{"charge_name": "Charge1", "details": "Some details"}]'
-        with patch("builtins.open", new_callable=mock_open, read_data=test_data):
-            result = self.cleaner.load_and_map_charge_names("fake_path.json")
-            self.assertEqual(result, {"Charge1": {"charge_name": "Charge1", "details": "Some details"}})
-
-        # Test empty file
-        with patch("builtins.open", new_callable=mock_open, read_data='[]'):
-            with self.assertRaises(FileNotFoundError):
-                self.cleaner.load_and_map_charge_names("empty.json")
-
-        # Test file not found
-        with patch("builtins.open", side_effect=FileNotFoundError):
-            with self.assertRaises(FileNotFoundError):
-                self.cleaner.load_and_map_charge_names("nonexistent.json")
-
-    def test_hash_defense_attorney(self):
-        input_data = {
-            "party information": {
-                "defense attorney": "John Doe",
-                "defense attorney phone number": "555-1234"
-            }
-        }
-        result = self.cleaner.hash_defense_attorney(input_data)
-        self.assertIsInstance(result, str)
-        self.assertNotEqual(result, "John Doe:555-1234")
-
-        # Test consistency
-        result2 = self.cleaner.hash_defense_attorney(input_data)
-        self.assertEqual(result, result2)
-
-        # Test different input
-        input_data2 = {
-            "party information": {
-                "defense attorney": "Jane Doe",
-                "defense attorney phone number": "555-5678"
-            }
-        }
-        result3 = self.cleaner.hash_defense_attorney(input_data2)
-        self.assertEqual(result, result3)
-
-        # Test missing data
-        input_data3 = {"party information": {}}
-        result4 = self.cleaner.hash_defense_attorney(input_data3)
-        self.assertEqual(result4, "")
-
-    def test_redact_cause_number(self):
-        # Test case 1: Normal input and consistency
-        input_dict = {"Case Metadata":{"code": "123-ABC-456"}}
-        result1 = self.cleaner.redact_cause_number(input_dict)
-        result2 = self.cleaner.redact_cause_number(input_dict)
-    
-        self.assertIsInstance(result1, str)
-        self.assertEqual(len(result1), 16)  # xxHash produces a 16-character hexadecimal string
-        self.assertEqual(result1, result2)  # Ensure consistent hashing
-    
-        # Test case 2: Different input produces different hash
-        input_dict2 = {"Case Metadata":{"code": "789-XYZ-012"}}
-        result3 = self.cleaner.redact_cause_number(input_dict2)
-        self.assertNotEqual(result1, result3)
-    
-        # Test case 3: Empty input
-        self.assertNotEqual(self.cleaner.redact_cause_number({"Case Metadata":{"code": ""}}), result1)
-    
-        # Test case 4: Missing 'code' key
-        with self.assertRaises(KeyError):
-            self.cleaner.redact_cause_number({})
-
-    def test_process_charges(self):
-        charges = [
-            {"level": "Misdemeanor", "charges": "Charge1", "statute": "123", "date": "12/01/2023"},
-            {"level": "Felony", "charges": "Charge2", "statute": "456", "date": "11/15/2023"},
-        ]
-        charge_mapping = {
-            "Charge1": {"mapped_field": "mapped_value1"},
-            "Charge2": {"mapped_field": "mapped_value2"}
-        }
-
-        processed_charges, earliest_date = self.cleaner.process_charges(charges, charge_mapping)
-
-        self.assertEqual(len(processed_charges), 2)
-        self.assertEqual(processed_charges[0]['charge_date'], "2023-12-01")
-        self.assertEqual(processed_charges[1]['charge_date'], "2023-11-15")
-        self.assertEqual(earliest_date, "2023-11-15")
-
-        # Test invalid date
-        charges_invalid_date = [{"level": "Misdemeanor", "charges": "Charge1", "statute": "123", "date": "invalid"}]
-        processed_charges, earliest_date = self.cleaner.process_charges(charges_invalid_date, charge_mapping)
-        self.assertEqual(len(processed_charges), 0)
-        self.assertEqual(earliest_date, "")
-    
-    def test_contains_good_motion(self):
-        self.assertTrue(self.cleaner.contains_good_motion("Motion To Suppress", "Event: Motion To Suppress"))
-        self.assertTrue(self.cleaner.contains_good_motion("Motion To Suppress", ["Other", "Motion To Suppress"]))
-        self.assertFalse(self.cleaner.contains_good_motion("Motion To Suppress", "Other Motion"))
-        self.assertFalse(self.cleaner.contains_good_motion("Motion To Suppress", ["Other1", "Other2"]))
-
-    def test_find_good_motions(self):
-        events = [
-            "Motion To Suppress",
-            "Motion to Reduce Bond",
-            "Other Event",
-            "Motion For Speedy Trial"
-        ]
-
-        result = self.cleaner.find_good_motions(events, cleaner.GOOD_MOTIONS)
-        self.assertEqual(len(result), 3)
-        self.assertEqual(result, ["Motion To Suppress", "Motion to Reduce Bond", "Motion For Speedy Trial"])
-
-        # Test with no matching motions
-        events_no_match = ["Other1", "Other2"]
-        result_no_match = self.cleaner.find_good_motions(events_no_match, cleaner.GOOD_MOTIONS)
-        self.assertEqual(result_no_match, [])
-
-    def test_process_single_case(self):
-        county = "hays"
-        input_folder_path = os.path.join(os.path.dirname(__file__), "..", "..", "resources", "test_files")
-        case_file = "test_123456.json"
-        output_folder_path = os.path.join(os.path.dirname(__file__), "..", "..", "resources", "test_files", "cleaned_test_json")
-
-        self.cleaner.process_single_case(input_folder_path, case_file, output_folder_path)
-
-        output_file_path = os.path.join(output_folder_path, case_file)
-
-        with open(output_file_path, 'r') as f:
-            output_data = json.load(f)
-            self.assertTrue("Case Metadata" in output_data)
-            self.assertTrue("Defendant Information" in output_data)
-            self.assertTrue("Charge Information" in output_data)
-            self.assertTrue("Case Details" in output_data)
-            self.assertTrue("parsing_date" in output_data)
-            self.assertTrue("html_hash" in output_data)
-            self.assertTrue("Good Motions" in output_data)
-            self.assertTrue("cause_number_redacted" in output_data)
-
-    # Will need 
-    """@patch("os.listdir", return_value=["case1.json", "case2.json"])
-    @patch("src.cleaner.Cleaner.get_or_create_folder_path")
-    @patch("src.cleaner.Cleaner.process_single_case")
-    def test_process_json_files(self, mock_process, mock_get_folder, mock_listdir):
-        county = "test_county"
-        folder_path = "case_json_folder"
-        mock_get_folder.return_value = "cleaned_folder_path"
-
-        self.cleaner.process_json_files(county, folder_path)
-
-        mock_get_folder.assert_called_once_with(county, "case_json_cleaned")
-        self.assertEqual(mock_process.call_count, 2)
-        mock_process.assert_any_call(folder_path, "case1.json", "cleaned_folder_path")
-        mock_process.assert_any_call(folder_path, "case2.json", "cleaned_folder_path")"""
-
-    @patch.object(cleaner.Cleaner, 'get_or_create_folder_path')
-    @patch.object(cleaner.Cleaner, 'process_json_files')
-    def test_clean(self, mock_process_json_files, mock_get_folder):
-        mock_get_folder.return_value = "mock_path"
-        county = "hays"
-
-        with self.assertLogs(level='INFO') as log:
-            self.cleaner.clean(county)
-
-        self.assertTrue(f"INFO:root:Processing data for county: {county}" in log.output)
-        self.assertTrue(f"INFO:root:Completed processing for county: {county}" in log.output)
-
-        mock_get_folder.assert_called_once_with(county, "case_json")
-        mock_process_json_files.assert_called_once_with(county, "mock_path")
-
-        # Test exception handling
-        mock_process_json_files.side_effect = Exception("Test error")
-        with self.assertLogs(level='ERROR') as log:
-            self.cleaner.clean(county)
-        self.assertIn(f"ERROR:root:Error during cleaning process for county: {county}. Error: Test error", log.output)
